Upgrade front-end dependencies.
Notice that there is a pending (Unresolved) serious security vulnerability in side our dependency path `favicons-webpack-plugin > favicons > to-ico > resize-img > jimp > url-regex`.

However, the effected module is only executed during the generation of favicon image&mainifests and will not effect the final executable. Thus, I'll not wait for the patch of the module this time, so we can ship other more important security and bug fixes without been delayed.