Added Server Message feature. Operator can now display a custom message on Home page of their Sshwifty via `ServerMessage` option of a Server. See README.md for detail
As an example, command:

    SSHWIFTY_LISTENPORT=8182 SSHWIFTY_SERVERMESSAGE="This service is hosted on [nirui.org](https://nirui.org)" ./sshwifty

will start a Sshwifty instance with a default server on TCP port 8182. On the Home page of the server, a text message will display which reads "This service is hosted on nirui.org" with "nirui.org" links to https://nirui.org.

If you configure Sshwifty via a JSON config file, there is a corresponding example added into `sshwifty.conf.example.json` for your reference.

Currently, Server Message only supports Link format in the form of `[text title](URL)`. The operator must ensure the correctness of their inputs, as bad format will either be sliently ignored or produce unpredictable result.