Allow HTTP server to output logs with proper logger
Still, due to how the HTTP server is designed, not all errors are logged.

For example, when Sshwifty is serving under plain HTTP mode, and the client sends a TLS handshake, no log will be added.

On the other hand, if Sshwifty is serving under TLS mode, and the client sends a plain HTTP handshake, a log will be added to indicate the fault.